Add default focus outline for focusable elements
We lost focus styles on certain controls in commit 4dad04ae3, since we
were applying a border with a rule which had more precedence than the
rule of border on focus.

We were already using a focus outline on links, but weren't doing the
same for other controls. Since sometimes browsers use a default outline
which is barely visible, this was very disorienting when browsing using
the keyboard; we were navigating through links that clearly indicated
where the keyboard focus was, and when reaching a button or form control
suddenly we had this almost imperceptible feedback. Even if I'm used to
it, my first reaction is always "where did the focus go?" until I
realize it's now on a button or form control.

Note that in the rules for the `:focus` styles we're including buttons
and inputs with the `[type]` attribute. This seems redundant since those
styles are already covered by the `button` and `input` selectors.
However, Foundation adds styles to buttons and inputs with the `[type]`
attribute. Since the attribute selector has precedence over the tag
selector, we need to use the attribute selector as well in order to
override Foundation's styles.